upgrade ExtUtils::MakeMaker on VMS
This version of PathTools removes long-deprecated VMS-related code
which had been used by EUMM until recently.  If we upgrade
PathTools, we might break EUMM, so we better upgrade it first.

The new EUMM is not a prereq of configuring PathTools, though, so we
don't need to make a configure_requires entry.